feat: Vite 4 + TS 5 + Vue 3.3 support
