--- vite.config.js
@@ -1,22 +1,0 @@
-import { defineConfig } from 'vite'
-import path from 'path';
-import vue from '@vitejs/plugin-vue';
-
-export default defineConfig({
-    plugins: [vue()],
-    build: {
-        lib: {
-            entry: path.resolve(__dirname, 'index.js'),
-            name: 'Vue3WebComponentWrapper',
-            fileName: (format) => `vue-web-component-wrapper.${format}.js`
-        },
-        rollupOptions: {
-            external: ['vue'],
-            output: {
-                globals: {
-                    vue: 'Vue'
-                }
-            }
-        }
-    }
-})
